+++ ft_type/ft_isdigit.c
@@ -1,20 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_tolower.c                                       :+:      :+:    :+:   */
+/*   ft_isdigit.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 16:10:44 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/02 16:16:42 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/02 15:31:31 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:44:32 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
-
-int	ft_tolower(int c)
+int	ft_isdigit(int c)
 {
-	if ('A' <= c && c <= 'Z')
-		return (c + 'a' - 'A');
-	return (c);
+	return ('0' <= c && c <= '9');
 }
