+++ ft_type/ft_tolower.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_isdigit.c                                       :+:      :+:    :+:   */
+/*   ft_tolower.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 15:31:31 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/02 15:31:53 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/02 16:10:44 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:44:37 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
-
-int	ft_isdigit(int c)
+int	ft_tolower(int c)
 {
-	return ('0' <= c && c <= '9');
+	if ('A' <= c && c <= 'Z')
+		return (c + 'a' - 'A');
+	return (c);
 }
