+++ ft_type/ft_toupper.c
@@ -6,11 +6,9 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/02 16:15:15 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/02 16:15:57 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:44:41 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
-#include "libft.h"
 
 int	ft_toupper(int c)
 {
