+++ include/ft_type.h
@@ -1,18 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_isprint.c                                       :+:      :+:    :+:   */
+/*   ft_type.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 16:08:11 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/02 16:17:17 by reasuke          ###   ########.fr       */
+/*   Created: 2024/02/24 12:44:55 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:45:31 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#ifndef FT_TYPE_H
+# define FT_TYPE_H
 
-int	ft_isprint(int c)
-{
-	return (' ' <= c && c <= '~');
-}
+int	ft_isalnum(int c);
+int	ft_isalpha(int c);
+int	ft_isascii(int c);
+int	ft_isdigit(int c);
+int	ft_isprint(int c);
+int	ft_tolower(int c);
+int	ft_toupper(int c);
+
+#endif
